refactor(common): storybook ui 문서와 ui package 분리
