+++ docs/storybook/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  root: true,
+  extends: ['@sickgyun/eslint-config/react-ts', 'plugin:storybook/recommended'],
+  ignorePatterns: ['*.js*', 'dist'],
+};
