+++ packages/ui/.eslintrc.js
@@ -1,5 +1,5 @@
 module.exports = {
   root: true,
-  extends: ['@sickgyun/eslint-config/react-ts', 'plugin:storybook/recommended'],
+  extends: ['@sickgyun/eslint-config/react-ts'],
   ignorePatterns: ['*.js*', 'dist'],
 };
