+++ pnpm-workspace.yaml
@@ -1,4 +1,5 @@
 packages:
   - 'apps/*'
-  - 'packages/*'
   - 'configs/*'
+  - 'docs/*'
+  - 'packages/*'
